adding report search, lock/unlock e2e tests
