+++ admin-frontend/e2e/pages/admin-portal-page.ts
@@ -1,6 +1,8 @@
 import { Page, expect } from '@playwright/test';
 import { User } from '../auth.setup';
 import { PagePaths } from '../utils';
+import { DateTimeFormatter, ZonedDateTime, ZoneId } from '@js-joda/core';
+import { Locale } from '@js-joda/locale_en';
 
 export class AdminPortalPage {
   constructor(public readonly page: Page) {}
@@ -27,7 +29,7 @@ export class AdminPortalPage {
     await expect(
       this.page.getByRole('link', { name: 'Announcements' }),
     ).toBeVisible();
-    // TODO: bring it back after the 
+    // TODO: bring it back after the
     // await expect(
     //   this.page.getByRole('link', { name: 'User Management' }),
     // ).toBeVisible();
@@ -52,5 +54,18 @@ export class AdminPortalPage {
     await this.page.getByRole('button', { name: 'Logout' }).click();
     await this.page.waitForTimeout(1000);
     await this.page.waitForURL(PagePaths.LOGOUT);
+  }
+
+  formatDate(
+    inDateStr: string,
+    inFormatter = DateTimeFormatter.ISO_DATE_TIME,
+    outFormatter = DateTimeFormatter.ofPattern('MMM d, yyyy').withLocale(
+      Locale.CANADA,
+    ),
+  ) {
+    const date = ZonedDateTime.parse(inDateStr, inFormatter);
+    const localTz = ZoneId.systemDefault();
+    const dateInLocalTz = date.withZoneSameInstant(localTz);
+    return outFormatter.format(dateInLocalTz);
   }
 }
